% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MinMod}
\alias{MinMod}
\title{Finds the minimal adequate model}
\usage{
MinMod(yX, test = "F", family = gaussian)
}
\arguments{
\item{yX}{A data frame with the dependent variable in the first column and the independent variables in the following columns}

\item{test}{"F" or "Chisq" for deletion testing and model comparison}

\item{family}{glm family}
}
\value{
Returns a list consisting of
     \item{ModelData}{Data frame containing the variables from the minimal adequate model}
     \item{BestModel}{glm object, best model}
}
\description{
Takes in a data frame and computes the minimal adequate model
}

