% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minmodelr.R
\docType{package}
\name{minmodelr}
\alias{minmodelr}
\alias{minmodelr-package}
\title{minmodelr: A (personal) package for automated deletion testing and single variable deletion testing}
\description{
The minmodelr package provides a functions for deletion testing according to Crawley
and on function for single variable deletion testing.
}
\section{minmodelr functions}{

\code{\link{MinMod}}
\code{\link{DelTestVar}}
}

