% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DelTestVar.R
\name{DelTestVar}
\alias{DelTestVar}
\title{Single variable deletion testing}
\usage{
DelTestVar(yX, family = "gaussian")
}
\arguments{
\item{yX}{A data frame with the dependent variable in the first column and the independent variables in the following columns}

\item{family}{Specifies glm type (e.g. "gaussian","binomial")}
}
\value{
Data frame with estimates, explained deviance, F-value, p-value (F-test), p-value (Chisq-test) per variable.
}
\description{
Tests for the significance of a single variable by deleting it from the model. The full model is then tested
against the reduced model with F-tests and Chisquared-tests. Explained deviance per variable is computed by substracting
the explained deviance from the reduced model (by ne variable) from the deviance explained by the full model.
}

